HHH-15912 adaptively choose the DDL type for ORDINAL enums based on the number of members
this amounts to a reversion of HHH-15288 for 99.99% of enums

+++ hibernate-core/src/test/java/org/hibernate/orm/test/annotations/enumerated/EnumeratedSmokeTest.java
@@ -79,7 +79,7 @@ public class EnumeratedSmokeTest extends BaseUnitTestCase {
 		assertThat( hibernateMappingEnumType.isOrdinal(), is(expectedJpaEnumType==EnumType.ORDINAL) );
 		final int expectedJdbcTypeCode = jdbcRegistry.getDescriptor(
 				expectedJpaEnumType == EnumType.ORDINAL ?
-						Types.SMALLINT :
+						Types.TINYINT :
 						Types.VARCHAR
 		).getJdbcTypeCode();
 		assertThat(
